requirements.txt
accelerate==0.20.3
albumentations==0.4.3
av==12.1.0
bezier==2023.7.28
bitsandbytes==0.38.1
black==24.4.2
cityscapesscripts==2.2.3
cloudpickle==2.2.1
docutils
einops==0.3.0
face_alignment==1.4.1
fairscale==0.4.13
ftfy==6.2.0
fvcore==0.1.5.post20221221
huggingface_hub==0.16.4
hydra-core==1.3.2
imageio==2.9.0
invisible_watermark==0.1.5
iopath==0.1.9
kornia==0.6.0
matplotlib==3.7.3
mmcv==2.2.0
mmdet==3.3.0
natsort==8.3.1
numpy==1.24.4
omegaconf==2.3.0
onnx==1.13.1
opencv_contrib_python==4.8.1.78
opencv_python==4.8.0.76
opencv_python_headless==4.7.0.72
pandas==1.5.3
psutil==5.9.5
pudb==2019.2
pycocotools==2.0.7
Pygments==2.14.0
pytest==8.2.2
pytorch_lightning==1.4.2
PyYAML==5.4.1
recommonmark==0.7.1
regex==2023.3.23
Requests==2.32.3
safetensors==0.3.3
scikit_learn==1.3.1
scipy==1.9.1
setuptools==65.6.3
Shapely==2.0.4
scikit-image
Sphinx
sphinx_rtd_theme==2.0.0
tabulate==0.9.0
taming_transformers==0.0.1
taming_transformers_rom1504==0.0.6
termcolor==2.4.0
timm==0.6.13
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.34.0
wandb==0.15.0
diffusers==0.21.0
Pillow==9.5.0
gdown